[UI] Behaviour change.
Left-click on system tray icon only brings the application on top (when minimized or in background) but doesn't keep it on top if looses focus.
Right-click on system tray icon brings the application on top if not minimized.
Minimizing application always sends it to system tray.
Incremented version from 0.28.0 to 0.29.0.