Refactor settings classes to use a new hierarchical layout.
Fixed an issue that was causing some settings not to be updated in the settings file when Enable checkboxes were not checked in the Option dialog.
Fixed an issue that was causing Polling Interval not to be updated in the settings file when WDT Sync was selected in the Device Configuration dialog.
Updated Options dialog validation messages.
Updated WindowPosition value for the default settings.
Allow user to input decimal values for the notification temperature threshold value.
Updated unit tests.