Refactor Temperature/Pressure/Radiation readings conversions.
Removed code that was dynamically updating notifications settings values configured by the user.
Fixed an issue within the Options dialog that was causing the Temperature threshold value not to be automatically converted when the measurement unit was changed.
Fixed an issue within the Options dialog that was causing Radiation threshold value to be displayed with one decimal digit when the measurement unit was set to cpm.